+++ CHANGELOG.md
@@ -4,7 +4,6 @@ Documentation of release versions of `nacc-attribute-deriver`
 
 ## Working (1.5.0.dev)
 
-* Adds rest of NP derived variables
 * Adds rest of UDS derived variables:
     * Form header
         * Here some NACC derived variables compute the same thing but apply to a different operation, so curation rule may point to a different derived variable's function
@@ -21,6 +20,8 @@ Documentation of release versions of `nacc-attribute-deriver`
     * D1 **except** NACCMCII (D1)
         * **NACCMCII is still not working properly, on hold for now**
 * Adds CLS derived variables
+* Adds MILESTONE derived variables
+* Adds rest of NP derived variables
 * Adds NACCNE4S and fixes NACCAPOE to account for ADC and ADGC genotypes not being the same
 * Refactors how dated values are handled - adds `dated` configuration to curation rules to be handled at curation rule level and removes redundant `_create` functions
     * Updates longtitudinal variables to be stored as list of `DateTaggedValue`s - curation rules updated